+++ options/options.go
@@ -95,7 +95,7 @@ var global *Options
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// Add adds a new supported option
+// Add adds a new option
 func (opts *Options) Add(name string, option *V) error {
 	if opts == nil {
 		return ErrOptionsIsNil
@@ -139,7 +139,7 @@ func (opts *Options) Add(name string, option *V) error {
 	return nil
 }
 
-// AddMap adds supported options as map
+// AddMap adds map with supported options
 func (opts *Options) AddMap(optMap Map) []error {
 	var errs []error
 
@@ -292,7 +292,31 @@ func (opts *Options) GetF(name string) float64 {
 	}
 }
 
-// Has check that option exists and set
+// Is checks if option with given name has given value
+func (opts *Options) Is(name string, value interface{}) bool {
+	if opts == nil {
+		return false
+	}
+
+	if !opts.Has(name) {
+		return false
+	}
+
+	switch t := value.(type) {
+	case string:
+		return opts.GetS(name) == t
+	case int:
+		return opts.GetI(name) == t
+	case float64:
+		return opts.GetF(name) == t
+	case bool:
+		return opts.GetB(name) == t
+	}
+
+	return false
+}
+
+// Has checks if option with given name exists and set
 func (opts *Options) Has(name string) bool {
 	if opts == nil {
 		return false
@@ -334,7 +358,7 @@ func (opts *Options) Parse(rawOpts []string, optMap ...Map) (Arguments, []error)
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// NewOptions create new options struct
+// NewOptions creates new options struct
 func NewOptions() *Options {
 	return &Options{
 		full:        make(Map),
@@ -343,7 +367,7 @@ func NewOptions() *Options {
 	}
 }
 
-// Add add new supported option
+// Add adds new supported option
 func Add(name string, opt *V) error {
 	if global == nil || !global.initialized {
 		global = NewOptions()
@@ -352,7 +376,7 @@ func Add(name string, opt *V) error {
 	return global.Add(name, opt)
 }
 
-// AddMap add supported option as map
+// AddMap adds map with supported options
 func AddMap(optMap Map) []error {
 	if global == nil || !global.initialized {
 		global = NewOptions()
@@ -397,13 +421,22 @@ func GetF(name string) float64 {
 	return global.GetF(name)
 }
 
-// Has check that option exists and set
+// Has checks if option with given name exists and set
 func Has(name string) bool {
 	if global == nil || !global.initialized {
 		return false
 	}
 
 	return global.Has(name)
+}
+
+// Is checks if option with given name has given value
+func Is(name string, value interface{}) bool {
+	if global == nil || !global.initialized {
+		return false
+	}
+
+	return global.Is(name, value)
 }
 
 // Parse parses slice with raw options
@@ -421,7 +454,7 @@ func ParseOptionName(name string) (string, string) {
 	return a.Long, a.Short
 }
 
-// Q merges several options to string
+// Q merges several options into string
 func Q(opts ...string) string {
 	return strings.Join(opts, " ")
 }